Basis writer - IMPORTANT NOTICE! READ BELOW:
This is enough for me already to close #2, but beware, this code has NOT BEEN TESTED AT ALL, meaning I cannot say at all if this works. Since we are talking about writing, it's always the more dangerous side of the job, eh? Anyway, I will test this all out later, and fix stuff as bugs arise, but I do recommend NOT to use the TJCRCreate and the TJCRCreateStream classes as long as long as I didn't proplerly test them out. I do not expact really destructive bugs, nor very spectecular ones, but still completely untested software remains completely untested software, so a warning is in order, so here goes